+++ R/birdie-package.R
@@ -14,7 +14,7 @@
 #' @importFrom RcppParallel CxxFlags
 #' @importFrom stats na.omit na.fail .lm.fit fitted simulate median cov
 #' @importFrom stats terms get_all_vars model.frame model.matrix setNames
-#' @importFrom stats formula update.formula as.formula nobs vcov residuals
+#' @importFrom stats formula update.formula as.formula nobs family vcov residuals
 #' @importFrom utils tail unstack relist head packageVersion
 #' @importFrom graphics barplot
 #' @importFrom grDevices colorRampPalette